+++ RedZone/blocks/Dispenser.java
@@ -141,8 +141,7 @@ public class Dispenser extends Block implements PoweredComponent
 		if (!w.isServer)
 		{
 			// System.out.printf("spawning new chest entity\n");
-			Entity eb = w.createEntityByName("RedZone:EntityDispenser", d, (float) (x) + 0.5f, (float) (y) + 0.5f,
-					(float) (z) + 0.5f);
+			Entity eb = w.createEntityByName("RedZone:EntityDispenser", d, x + 0.5, y + 0.5, z + 0.5);
 			if (eb != null)
 			{
 				eb.init();
@@ -184,8 +183,7 @@ public class Dispenser extends Block implements PoweredComponent
 		if (!w.isServer)
 		{
 			// System.out.printf("onBlockPlaced spawning new dispenser entity\n");
-			Entity eb = w.createEntityByName("RedZone:EntityDispenser", dimension, (float) (x) + 0.5f,
-					(float) (y) + 0.5f, (float) (z) + 0.5f);
+			Entity eb = w.createEntityByName("RedZone:EntityDispenser", dimension, x + 0.5, y + 0.5, z + 0.5);
 			if (eb != null)
 			{
 				eb.init();
