+++ RedZone/blocks/PusherCornerPipe.java
@@ -90,8 +90,7 @@ public class PusherCornerPipe extends CornerPipe implements PoweredComponent
 			if (w.isServer)
 			{
 				// System.out.printf("spawning new chest entity\n");
-				Entity eb = w.createEntityByName("RedZone:EntityPusherCornerPipe", d, (float) (x) + 0.5f,
-						(float) (y) + 0.5f, (float) (z) + 0.5f);
+				Entity eb = w.createEntityByName("RedZone:EntityPusherCornerPipe", d, x + 0.5, y + 0.5, z + 0.5);
 				if (eb != null)
 				{
 					eb.init();
@@ -152,8 +151,7 @@ public class PusherCornerPipe extends CornerPipe implements PoweredComponent
 		if (w.isServer)
 		{
 			// System.out.printf("onBlockPlaced spawning new dispenser entity\n");
-			Entity eb = w.createEntityByName("RedZone:EntityPusherCornerPipe", dimension, (float) (x) + 0.5f,
-					(float) (y) + 0.5f, (float) (z) + 0.5f);
+			Entity eb = w.createEntityByName("RedZone:EntityPusherCornerPipe", dimension, x + 0.5, y + 0.5, z + 0.5);
 			if (eb != null)
 			{
 				eb.init();
