+++ RedZone/blocks/TractorShooter.java
@@ -19,8 +19,19 @@ import mechanics.Orienter;
 import mechanics.PoweredComponent;
 
 /**
+ * Tractor shooters shoot out tractor beams, which can move entities and
+ * blocks around.
+ * <p>
+ * Tractor shooters, when powered, shoot a tractor beam out of their front.
+ * How far this beam goes is determined by the closest entity or block in that
+ * direction, or the power supplied, whichever is smaller. If the tractor beam
+ * is not obstructed, it will cause the first block or entity it hits to move
+ * towards the tractor shooter.
+ * 
  * @author eaglgenes101
- *
+ * @see TractorShooter
+ * @see RepulsorBeam
+ * @see EntityPushedBlock
  */
 public class TractorShooter extends Block implements PoweredComponent
 {
@@ -90,7 +101,7 @@ public class TractorShooter extends Block implements PoweredComponent
 		{
 			
 			List<Entity> nearby_list = DangerZone.entityManager.findEntitiesInRange(2.0f, d,
-					x + offset[0] * reach + 0.5f, y + offset[1] * reach + 0.5f, z + offset[2] * reach + 0.5f);
+					x + offset[0] * reach + 0.5, y + offset[1] * reach + 0.5, z + offset[2] * reach + 0.5);
 			ListIterator<Entity> li;
 
 			if (nearby_list != null)
@@ -110,15 +121,15 @@ public class TractorShooter extends Block implements PoweredComponent
 						if (e instanceof EntityLiving)
 						{
 							EntityLiving el = (EntityLiving) e;
-							int intheight = (int) (el.height + 0.995f);
-							float dx, dz;
+							int intheight = (int) (el.height + 0.995);
+							double dx, dz;
 							for (int k = 0; k < intheight; k++)
 							{
 								if ((int) el.posy + k == y + offset[1] * reach)
 								{
-									dx = el.posx - ((float) x + offset[0] * reach + 0.5f);
-									dz = el.posz - ((float) z + offset[2] * reach + 0.5f);
-									if (Math.sqrt((dx * dx) + (dz * dz)) < (0.5f + (el.width / 2.0f)))
+									dx = el.posx - (x + offset[0] * reach + 0.5);
+									dz = el.posz - (z + offset[2] * reach + 0.5);
+									if (Math.sqrt((dx * dx) + (dz * dz)) < (0.5 + (el.width / 2.0)))
 									{
 										break outerLoop;
 									}
